Better MD5 mismatch error description
The wording for this md5 mismatch error was formerly very sparse. Some people using this to install wordpress may not know what MD5 hashes and comparisons are, so  a slight description with a suggestion as to what to look into to cure the error, have been added. This could possibly occur if say wordpress.org changed out their site structure or rarely if some hacker man in the middle attack occurred.